refactor: add abstract for DefaultMessage & KafkaMessage to manage attributes & internalAttributes
