+++ src/test/java/io/gravitee/gateway/reactive/api/message/DefaultMessageTest.java
@@ -142,7 +142,7 @@ class DefaultMessageTest {
 
     @Test
     void shouldHaveCorrelationIdFromBufferConstructor() {
-        DefaultMessage cut = new DefaultMessage(null);
+        DefaultMessage cut = new DefaultMessage((String) null);
         assertNotNull(cut.correlationId());
     }
